requirements.txt
#LangChain Core
langchain
langchain-core
langchain_groq
langchain_community
# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
# Environment Variable Management
python-dotenv
# Machine Learning Utilities
numpy
scikit-learn
# PDF Processing
pypdf
# Streamlit
streamlit
#Other
faiss-cpu
sentence-transformers
fastapi
uvicorn